Update main.js
Changed things up so that the game will go into the canvas that I have made for it. Sometimes I think way to hard about these things. Like... Am I a god to the 1s and 0s that are arranged into the canvas? Do they await the day that I will deliver them a miracle? Do they worship me?